Basic Functionality
The code seems to work. I haven't tested it very thoroughly though and the localSearch function desperately needs to be optimized. There also seems to be an issue where the code population becomes filled with all of the same codes after a few generations. So have to find a way to maintain diversity across generations. Maybe this is an artifact of how the localSearch function is implemented.